+++ apps/nutrition-tracker/feature-macros/src/main/java/com/personalapps/suite/nutrition/feature/macros/domain/usecase/SaveMacroGoalUseCase.kt
@@ -5,7 +5,7 @@ import com.personalapps.suite.nutrition.feature.api.repository.MacroGoalReposito
 import com.personalapps.suite.shared.common.Result
 
 class SaveMacroGoalUseCase(private val repository: MacroGoalRepository) {
-    suspend operator fun invoke(calories: Int, protein: Float, carbs: Float, fat: Float): Result<Long> {
+    suspend operator fun invoke(calories: Int, protein: Int, carbs: Int, fat: Int): Result<Long> {
         if (calories <= 0) {
             return Result.Error(IllegalArgumentException("Calories must be greater than 0"))
         }
